# docs/openapi/stock-ai.openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Stock-AI API Contract (Snapshot)",
    "version": "v1.1.0",
    "description": "Snapshot contract for typed client generation. Refresh with scripts/openapi_sync.ps1."
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Local backend"
    }
  ],
  "paths": {
    "/api/auth/register": {
      "post": {
        "tags": ["auth"],
        "summary": "Register user",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/login": {
      "post": {
        "tags": ["auth"],
        "summary": "Login user",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/analysis": {
      "post": {
        "tags": ["analysis"],
        "summary": "Run stock analysis",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnalysisRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalysisResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/analysis/{id}": {
      "get": {
        "tags": ["analysis"],
        "summary": "Get analysis detail",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true
                }
              }
            }
          }
        }
      }
    },
    "/api/portfolio": {
      "post": {
        "tags": ["portfolio"],
        "summary": "Create portfolio",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePortfolioRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PortfolioResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/portfolio/{id}/position": {
      "post": {
        "tags": ["portfolio"],
        "summary": "Add position",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddPositionRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "integer",
                    "format": "int64"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/portfolio/{id}/risk-dashboard": {
      "get": {
        "tags": ["portfolio"],
        "summary": "Get risk dashboard",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "lookbackDays",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 252
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RiskDashboardResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/portfolio/{id}/earnings-calendar": {
      "get": {
        "tags": ["portfolio"],
        "summary": "Get earnings calendar",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "daysAhead",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 90
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EarningsCalendarResponse"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "RegisterRequest": {
        "type": "object",
        "required": ["email", "password"],
        "properties": {
          "email": {
            "type": "string",
            "format": "email"
          },
          "password": {
            "type": "string",
            "minLength": 8
          }
        }
      },
      "LoginRequest": {
        "type": "object",
        "required": ["email", "password"],
        "properties": {
          "email": {
            "type": "string",
            "format": "email"
          },
          "password": {
            "type": "string"
          }
        }
      },
      "AuthResponse": {
        "type": "object",
        "properties": {
          "accessToken": {
            "type": "string"
          },
          "refreshToken": {
            "type": "string"
          },
          "requires2FA": {
            "type": "boolean"
          },
          "pendingToken": {
            "type": "string"
          }
        },
        "additionalProperties": true
      },
      "AnalysisRequest": {
        "type": "object",
        "required": ["ticker", "market", "horizonDays", "riskProfile"],
        "properties": {
          "ticker": {
            "type": "string"
          },
          "market": {
            "type": "string",
            "enum": ["US", "KR"]
          },
          "horizonDays": {
            "type": "integer"
          },
          "riskProfile": {
            "type": "string",
            "enum": ["conservative", "moderate", "aggressive"]
          }
        }
      },
      "AnalysisResponse": {
        "type": "object",
        "properties": {
          "runId": {
            "type": "integer",
            "format": "int64"
          },
          "result": {
            "type": "object",
            "additionalProperties": true
          }
        },
        "additionalProperties": true
      },
      "CreatePortfolioRequest": {
        "type": "object",
        "required": ["name", "riskProfile"],
        "properties": {
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "targetReturn": {
            "type": "number"
          },
          "riskProfile": {
            "type": "string",
            "enum": ["conservative", "moderate", "aggressive"]
          }
        }
      },
      "PortfolioResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "type": "string"
          }
        },
        "additionalProperties": true
      },
      "AddPositionRequest": {
        "type": "object",
        "required": ["ticker", "market", "quantity", "entryPrice"],
        "properties": {
          "ticker": {
            "type": "string"
          },
          "market": {
            "type": "string",
            "enum": ["US", "KR"]
          },
          "quantity": {
            "type": "number"
          },
          "entryPrice": {
            "type": "number"
          },
          "entryDate": {
            "type": "string",
            "format": "date"
          },
          "notes": {
            "type": "string"
          }
        }
      },
      "RiskDashboardResponse": {
        "type": "object",
        "properties": {
          "portfolioId": {
            "type": "integer",
            "format": "int64"
          },
          "metrics": {
            "type": "object",
            "additionalProperties": true
          },
          "positions": {
            "type": "array",
            "items": {
              "type": "object",
              "additionalProperties": true
            }
          }
        },
        "additionalProperties": true
      },
      "EarningsCalendarResponse": {
        "type": "object",
        "properties": {
          "portfolioId": {
            "type": "integer",
            "format": "int64"
          },
          "items": {
            "type": "array",
            "items": {
              "type": "object",
              "additionalProperties": true
            }
          }
        },
        "additionalProperties": true
      }
    }
  }
}